+++ examples/mp4_to_h264.cpp
@@ -6,6 +6,7 @@
 #include <iostream>
 #include <fstream>
 #include <string>
+#include <vector>
 
 #include <petro/extractor/nalu_extractor.hpp>