+++ NEWS.rst
@@ -6,7 +6,8 @@ every change, see the Git log.
 
 Latest
 ------
-* tbd
+* Minor: Changed `avc_sample_extractor` to use the pimpl pattern. 
+* Minor: Enabled building petro as a shared library
 
 5.0.0
 -----